package/toolmate_before_2_0/requirements.txt
# common
numpy==1.26.4
pandas==2.2.3
protobuf==4.25.3
ollama==0.4.7
groq==0.15.0
mistralai==1.4.0
openai==1.60.0
tiktoken==0.8.0
anthropic==0.45.0
google-genai==0.6.0
autogen[anthropic,gemini,mistral,groq,ollama,retrievechat,teachable,lmm,graph,long-context]==0.5.3
autogenstudio==0.1.5
langchain-unstructured[local]==0.1.6
unstructured[all-docs]==0.15.14
unstructured-client==0.28.1
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.13
litellm[proxy]==1.54.1
huggingface-hub==0.27.1
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-openai==0.3.2
langchain-huggingface==0.1.2
langchain-ollama==0.2.2
langchain-text-splitters==0.3.5
langsmith==0.3.1
chromadb==0.5.3
chroma-hnswlib==0.7.3
sentence-transformers==3.4.0
scikit-image==0.24.0
scikit-learn==1.6.0
kiwisolver==1.4.7
termcolor==2.5.0
wcwidth==0.2.13
aiosignal==1.3.1
aiohappyeyeballs==2.4.4
jsonpath-python==1.0.6
annotated-types==0.7.0
exceptiongroup==1.2.2
h11==0.14.0
cachetools==5.5.0
marshmallow==3.23.1
httplib2==0.22.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.75.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.62.3
cryptography==42.0.8
sqlmodel==0.0.22
tabulate==0.9.0
xlrd==2.0.1
humanfriendly==10.0
httptools==0.6.4
joblib==1.4.2
opentelemetry_sdk==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-api==1.27.0
matplotlib==3.9.3
matplotlib-inline==0.1.7
onnxruntime==1.19.2
importlib_metadata==8.4.0
jedi==0.19.2
MarkupSafe==3.0.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
# more general
wikipedia==1.4.0
xyzservices==2024.9.0
zope.interface==7.0.3
blinker==1.9.0
deepdiff==8.0.1
itsdangerous==2.2.0
imageio==2.35.1
multitasking==0.0.11
llvmlite==0.43.0
lazy-loader==0.4
#llama-cpp-python[server]==0.3.0
cmake==3.30.3
#transformers==4.44.2
soundfile==0.12.1
sounddevice==0.5.0
elevenlabs==1.5.0
#guidance==0.1.16
# no version restriction
wheel
setuptools
pipdeptree
googlesearch-python
google-cloud-speech
google-cloud-texttospeech
requests
argparse
pendulum
folium
seaborn[stats]
sympy
prompt_toolkit
Pygments
datetime
#netifaces
geocoder
art
apsw
gTTS
yt-dlp
rembg
qrcode
pyperclip
colorama
pillow
docker
einops
python-magic
pygame
#PySide6
feedparser
html2text
pypdf
PyMuPDF
yfinance
setuptools-rust
SpeechRecognition
openai-whisper
pydub
#stable-diffusion-cpp-python
pytz
geopy
psutil
flashtext
#flaml[automl]
tavily-python
wikipedia
tqdm
vosk
edge-tts
ftfy
lightrag-hku
fastapi
uvicorn
xonsh[full]
markitdown
pyautogui
# Captain Agent requirements
# those are commented are duplicates
#markdownify
#arxiv
pymupdf
wikipedia-api
easyocr
#python-pptx
#openai-whisper
#pandas
#scipy
#sentence-transformers